Fix or silence some compiler warnings about narrowing conversions, unused variables, name hiding and other minor stuff

+++ test/test_arithmetic.hpp
@@ -1100,14 +1100,14 @@ void test_float_ops(const boost::mpl::int_<boost::multiprecision::number_kind_fl
       BOOST_CHECK_EQUAL(scalbn(Real(2), 5), 2 * pow(double(std::numeric_limits<Real>::radix), 5));
       BOOST_CHECK_EQUAL(scalbn(Real(2), -5), Real(2) / pow(double(std::numeric_limits<Real>::radix), 5));
       v = 512;
-      exponent = ilogb(v);
+      exponent = static_cast<int>(ilogb(v));
       r = scalbn(v, -exponent);
       BOOST_CHECK(r >= 1);
       BOOST_CHECK(r < std::numeric_limits<Real>::radix);
       BOOST_CHECK_EQUAL(exponent, logb(v));
       BOOST_CHECK_EQUAL(v, scalbn(r, exponent));
       v = 1 / v;
-      exponent = ilogb(v);
+      exponent = static_cast<int>(ilogb(v));
       r = scalbn(v, -exponent);
       BOOST_CHECK(r >= 1);
       BOOST_CHECK(r < std::numeric_limits<Real>::radix);
